update functionality remaining to be implemented.

+++ POSStore/MainWindow.xaml.cs
@@ -24,21 +24,23 @@ namespace POSStore
     public partial class MainWindow : Window
     {
         //public string connectionString = "Data Source=ENG-RNR-05;Initial Catalog = DSPOS; Integrated Security = True";
-        public string connectionString = "Data Source=AHSAN-PC\\SQLExpress;Initial Catalog=DSPOS;Integrated Security=True;Pooling=False";
+        public string? connectionString = "Data Source=AHSAN-PC\\SQLExpress;Initial Catalog=DSPOS;Integrated Security=True;Pooling=False";
         public string queryString = "SELECT * from mainLedger";
-        public SqlConnection connection;
-        public SqlDataAdapter dataAdapter;
+        public SqlConnection? connection;
+        public SqlDataAdapter? dataAdapter;
         public MainWindow()
         {
             InitializeComponent();
 
             DataTable dTable = new DataTable();
-            try {
-            connection = new SqlConnection(connectionString);
-            dataAdapter = new SqlDataAdapter(queryString, connectionString);
-            dataAdapter.Fill(dTable);
-            connection.Close();
-            } catch (Exception e)
+            try
+            {
+                connection = new SqlConnection(connectionString);
+                dataAdapter = new SqlDataAdapter(queryString, connectionString);
+                dataAdapter.Fill(dTable);
+                connection.Close();
+            }
+            catch (Exception e)
             {
                 MessageBox.Show("Connot Load Data from SQL database" +
                  Environment.NewLine + e.Message +
@@ -96,9 +98,26 @@ namespace POSStore
             dataAdapter.Fill(dTable);
             drugLedger.DataContext = null;
             drugLedger.DataContext = dTable.DefaultView;
-            connection.Close();
+            // connection.Close();
         }
+        private bool validateEntry()
+        {
+            /// this function is to vertify that user
+            /// is not selecting invalid entry
+            /// only check if first column of table is filled
+            bool valid = false;
+            List<DataGridCellInfo> cells = drugLedger.SelectedCells.ToList();
+            string idData = string.Empty;
+            var cellContent = cells[0].Column.GetCellContent(cells[0].Item);
+            TextBlock cellText = (TextBlock)cellContent;
+            idData = cellText.Text.Trim();
+            if (!string.IsNullOrEmpty(idData))
+            {
+                return true;
+            }
+            return valid;
 
+        }
         private void viewEntry(object sender, RoutedEventArgs e)
         {
             ///getting a data from the datagrid cell is very tricky
@@ -141,16 +160,46 @@ namespace POSStore
 
         private void deleteEntry(object sender, RoutedEventArgs e)
         {
-            DataRowView row = drugLedger.SelectedItem as DataRowView;
-            string cString = @"DELETE FROM mainLedger WHERE id='" + row[0].ToString() + "';";
-            //MessageBox.Show(cString);
-            addDatatoTable(cString);
+            if (validateEntry())
+            {
+                DataRowView? row = drugLedger.SelectedItem as DataRowView;
+                string cString = @"DELETE FROM mainLedger WHERE id='" + row[0].ToString() + "';";
+                //MessageBox.Show(cString);
+                addDatatoTable(cString);
+            }
         }
 
-        public void viewDrugData(object sender, RoutedEventArgs e)
+        public void viewDrugData(object sender, RoutedEventArgs evt)
         {
-            drugView dv = new drugView();
-            dv.ShowDialog();
+            if (validateEntry())
+            {
+                DataRowView? row = drugLedger.SelectedItem as DataRowView;
+                string idQuery = @"SELECT * FROM mainLedger WHERE id='" + row[0].ToString() + "';";
+                DataTable dTable = new DataTable();
+                try
+                {
+                    connection = new SqlConnection(connectionString);
+                    dataAdapter = new SqlDataAdapter(idQuery, connectionString);
+                    dataAdapter.Fill(dTable);
+                    connection.Close();
+                }
+                catch (Exception e)
+                {
+                    MessageBox.Show("valid data error" +
+                     Environment.NewLine + e.Message +
+                     Environment.NewLine + e.Source,
+                     "Error",
+                     MessageBoxButton.OK,
+                     MessageBoxImage.Error
+                     );
+                }
+                drugView dv = new drugView(dTable);
+                dv.ShowDialog();
+            }
+        }
+        public void addDrugData(object sender, RoutedEventArgs evt)
+        {
+            viewEntry(this,new RoutedEventArgs());
         }
     }
 }
